fabric-finance/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
python-dotenv==1.0.1
httpx==0.27.0
python-telegram-bot==21.3
apscheduler==3.10.4
pydantic==2.7.1
python-dateutil==2.9.0
aiofiles==23.2.1
